fix: authValidation bug
