+++ app/backend/src/app.ts
@@ -1,11 +1,13 @@
 import express from 'express';
 import 'express-async-errors';
-import {
-  userRoutes,
-  authRoutes,
-  productRoutes,
-} from './routes';
 import errorHandler from './middlewares/errorHandler';
+import {
+  authRoutes,
+  publicProductRoutes,
+  privateProductRoutes,
+  userRoutes,
+} from './routes';
+import authValidation from './middlewares/authValidation';
 
 class App {
   public app: express.Express;
@@ -22,7 +24,7 @@ class App {
     const accessControl: express.RequestHandler = (_req, res, next) => {
       res.header('Access-Control-Allow-Origin', '*');
       res.header('Access-Control-Allow-Methods', 'GET,POST,DELETE,OPTIONS,PUT,PATCH');
-      res.header('Access-Control-Allow-Headers', '*');
+      res.header('Access-Control-Allow-Headers', ['Content-Type', 'Authorization']);
       next();
     };
 
@@ -30,7 +32,9 @@ class App {
     this.app.use(accessControl);
     this.app.use('/users', userRoutes);
     this.app.use('/auth', authRoutes);
-    this.app.use('/products', productRoutes);
+    this.app.use('/products', publicProductRoutes);
+    this.app.use(authValidation);
+    this.app.use('/products', privateProductRoutes);
     this.app.use(errorHandler);
   }
